refactor: rename lua annotation generator module
